Reservation page 70% finnish

+++ lib/reservation_details.dart
@@ -23,8 +23,31 @@ class ReservationDetails extends StatelessWidget {
             const SizedBox(height: 30), // Adjust the height as needed
             ElevatedButton(
               onPressed: () {
-                onDelete();
-                Navigator.pop(context);
+                showDialog(
+                  context: context,
+                  builder: (BuildContext context) {
+                    return AlertDialog(
+                      title: Text(language[28]),
+                      content: Text(language[29]),
+                      actions: <Widget>[
+                        TextButton(
+                          onPressed: () {
+                            Navigator.of(context).pop(); // Dismiss the dialog
+                          },
+                          child: Text(language[30]),
+                        ),
+                        TextButton(
+                          onPressed: () {
+                            Navigator.of(context).pop();
+                            Navigator.pop(context);
+                            onDelete();
+                          },
+                          child: Text(language[31]),
+                        ),
+                      ],
+                    );
+                  },
+                );
               },
               child: Text(language[6]),
             ),
